requirements-dev.txt
# These requirements are for development and test of the library only

# Library dependencies should be specified in setup.py with as wide a
# version range as possible, for maximum compatibility with other tools

# Testing
# ------------------------------------------------------------------------------
tox
pluggy
gcloud-storage-emulator==0.3.2
google-cloud-storage>=1.35.1

# Code quality
# ------------------------------------------------------------------------------
flake8  # https://github.com/PyCQA/flake8
black  # https://github.com/ambv/black
pre-commit  # https://github.com/pre-commit/pre-commit
coverage

# Pre-deploy checks
# ------------------------------------------------------------------------------
setuptools
wheel
twine  # <---- nothing to do with the twined library!

# Current library (editable install)
#    Installs any dependencies in setup.py
# ------------------------------------------------------------------------------
-e .
